BugFix: Loop de Echo infinito causado pelo ultimo commit
This commit introduces filtering at the beginning of the `SystemNotificationManagerImpl.processNotification()` method using `SystemNotificationValidator.validNotification()`.

The new `SystemNotificationValidator.validNotification()` method checks if a `StatusBarNotification` is valid based on the criteria defined in `applyDefaultFilter()`:
- Not from the current application.
- Has either a title or content.
- Is unique by title and content.
- Is unique by key.

This ensures that only valid notifications proceed to further processing, potentially reducing unnecessary operations for notifications that would be filtered out later.

Signed-off-by: Gilian Marques <gilian762@gmail.com>