Refactor: Improve logic for "Post Now" option in context menu
The condition for showing the "Post Now" option in the notification context menu has been simplified.

Previously, the condition was `(!not.permaHidden && not.isSnoozed) || not.isSystemSnoozed`.
It has been changed to `!not.permaHidden && (not.isSnoozed || not.isSystemSnoozed)`.

This change ensures that the "Post Now" option is shown if the notification is not permanently hidden AND it is either user-snoozed or system-snoozed.

Signed-off-by: Gilian Marques Fernandes <gilian762@gmail.com>