Refactor: Adjust timeout for NotificationListener access
The `PostAppSnoozedNotificationsUseCase` has been updated to use the new `NotificationListener.getWhenReady(timeoutMillis)` method. The timeout for obtaining the `NotificationListener` service has been adjusted from 500ms within a `withTimeoutOrNull` block to a direct 500ms timeout parameter in the `getWhenReady` call.

This change streamlines the process of accessing the notification listener service while maintaining the timeout mechanism to prevent the app from hanging if notification access permission is not granted.

Signed-off-by: Gilian Marques Fernandes <gilian762@gmail.com>