Refactor: Remove redundant `AppLogger.d()` overload
This commit removes the no-argument `AppLogger.d()` overload as it was redundant. The `AppLogger.d(msg: String, vararg relevantObjects: Any)` method can be called with an empty string to achieve the same result.

Calls to the removed overload in `RescheduleReportNotificationsOnBootUseCase.kt` and `NotificationListener.kt` have been updated to `AppLogger.d("")`.

A minor formatting change (newline) was made in `SystemNotificationManagerImpl.kt`.

Signed-off-by: Gilian Marques Fernandes <gilian762@gmail.com>